Fixed the number to reflext amount of alert headers
