+++ app/component/AlertList.js
@@ -1,51 +1,13 @@
 import cx from 'classnames';
 import connectToStores from 'fluxible-addons-react/connectToStores';
-import isEmpty from 'lodash/isEmpty';
-import groupBy from 'lodash/groupBy';
-import uniqBy from 'lodash/uniqBy';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 
 import ComponentUsageExample from './ComponentUsageExample';
 import RouteAlertsRow from './RouteAlertsRow';
-import { isAlertValid } from '../util/alertUtils';
-import { routeNameCompare } from '../util/searchUtils';
+import { alertCompare, createUniqueAlertList } from '../util/alertUtils';
 import { AlertSeverityLevelType } from '../constants';
-
-/**
- * Compares the given alerts in order to sort them.
- *
- * @param {*} a the first alert to compare.
- * @param {*} b the second alert to compare.
- */
-export const alertCompare = (a, b) => {
-  // sort by expiration status
-  if (a.expired !== b.expired) {
-    return a.expired ? 1 : -1;
-  }
-
-  // sort by missing route information (for stop level alerts)
-  if (!a.route || !a.route.shortName) {
-    // sort by stop information if it exists
-    if (a.stop && b.stop) {
-      return `${a.stop.code}`.localeCompare(`${b.stop.code}`);
-    }
-    return -1;
-  }
-
-  // sort by route information
-  const routeOrder = routeNameCompare(a.route || {}, b.route || {});
-  if (routeOrder !== 0) {
-    return routeOrder;
-  }
-
-  // sort by alert validity period
-  return b.validityPeriod.startTime - a.validityPeriod.startTime;
-};
-
-const hasRoute = alert => alert && !isEmpty(alert.route);
-const hasStop = alert => alert && !isEmpty(alert.stop);
 
 const AlertList = ({
   cancelations,
@@ -55,50 +17,15 @@ const AlertList = ({
   serviceAlerts,
   showRouteNameLink,
 }) => {
-  const getRoute = alert => alert.route || {};
-  const getMode = alert => getRoute(alert).mode;
-  const getShortName = alert => getRoute(alert).shortName;
-  const getRouteGtfsId = alert => getRoute(alert).gtfsId;
+  const groupedAlerts =
+    createUniqueAlertList(
+      serviceAlerts,
+      cancelations,
+      currentTime,
+      showExpired,
+    ) || [];
 
-  const getStop = alert => alert.stop || {};
-  const getVehicleMode = alert => getStop(alert).vehicleMode;
-  const getCode = alert => getStop(alert).code;
-  const getStopGtfsId = alert => getStop(alert).gtfsId;
-
-  const getGroupKey = alert =>
-    `${alert.severityLevel}${(hasRoute(alert) && `route_${getMode(alert)}`) ||
-      (hasStop(alert) && `stop_${getVehicleMode(alert)}`)}${alert.header}${
-      alert.description
-    }`;
-  const getUniqueId = alert =>
-    `${getShortName(alert) || getCode(alert)}${getGroupKey(alert)}`;
-
-  const uniqueAlerts = uniqBy(
-    [
-      ...(Array.isArray(cancelations)
-        ? cancelations
-            .map(cancelation => ({
-              ...cancelation,
-              severityLevel: AlertSeverityLevelType.Warning,
-              expired: !isAlertValid(cancelation, currentTime, {
-                isFutureValid: true,
-              }),
-            }))
-            .filter(alert => (showExpired ? true : !alert.expired))
-        : []),
-      ...(Array.isArray(serviceAlerts)
-        ? serviceAlerts
-            .map(alert => ({
-              ...alert,
-              expired: !isAlertValid(alert, currentTime),
-            }))
-            .filter(alert => (showExpired ? true : !alert.expired))
-        : []),
-    ],
-    getUniqueId,
-  );
-
-  if (uniqueAlerts.length === 0) {
+  if (groupedAlerts.length === 0) {
     return (
       <div className="stop-no-alerts-container">
         <FormattedMessage
@@ -108,41 +35,6 @@ const AlertList = ({
       </div>
     );
   }
-
-  const alertGroups = groupBy(uniqueAlerts, getGroupKey);
-  const groupedAlerts = Object.keys(alertGroups).map(key => {
-    const alerts = alertGroups[key];
-    const alert = alerts[0];
-    return {
-      ...alert,
-      route:
-        (hasRoute(alert) && {
-          mode: getMode(alert),
-          routeGtfsId: alerts
-            .sort(alertCompare)
-            .map(getRouteGtfsId)
-            .join(','),
-          shortName: alerts
-            .sort(alertCompare)
-            .map(getShortName)
-            .join(', '),
-        }) ||
-        undefined,
-      stop:
-        (hasStop(alert) && {
-          stopGtfsId: alerts
-            .sort(alertCompare)
-            .map(getStopGtfsId)
-            .join(','),
-          code: alerts
-            .sort(alertCompare)
-            .map(getCode)
-            .join(', '),
-          vehicleMode: getVehicleMode(alert),
-        }) ||
-        undefined,
-    };
-  });
   return (
     <div className={cx({ 'momentum-scroll': !disableScrolling })}>
       <div className="route-alerts-list">
